unsigned longs for time

+++ src/platforms/hosted.c
@@ -280,12 +280,12 @@ void *platform_interrupt_thread(void *_interrupt_fd) {
     char output[64];
     switch (msg.type) {
       case TRIGGER0:
-        sprintf(output, "# TRIGGER0 %d\n", msg.time);
+        sprintf(output, "# TRIGGER0 %lu\n", (unsigned long)msg.time);
         write(STDERR_FILENO, output, strlen(output));
         decoder_update_scheduling(&(struct decoder_event){.trigger = 0, .time = msg.time}, 1);
         break;
       case TRIGGER1:
-        sprintf(output, "# TRIGGER1 %d\n", msg.time);
+        sprintf(output, "# TRIGGER1 %lu\n", (unsigned long)msg.time);
         write(STDERR_FILENO, output, strlen(output));
         decoder_update_scheduling(&(struct decoder_event){.trigger = 1, .time = msg.time}, 1);
         break;
